pyproject.toml
[tool.poetry]
name = "Conformalized Xplainable Malaria"
version = "0.0.1"
description = " "
authors = ["Mateus Balda <mateusbalda89@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<=3.12.7"
numpy = "^2"
pandas = "^2.3.3"
matplotlib = "^3.10.6"
seaborn = "^0.13.2"
scikit-learn = "^1.7.2"
plotly = "^6.3.1"
scipy = "^1.16.2"
statsmodels = "^0.14.5"
xgboost = "^3.1.1"
shap = "^0.49.1"
imblearn = "^0.0"
tensorflow = "^2.20.0"
opencv-python = "^4.12.0.88"
pillow = "^12.0.0"
torch = "^2.9.1"
torchvision = "^0.24.1"
torchaudio = "^2.9.1"
ultralytics = "^8.3.235"
grad-cam = "^1.5.5"
crepes = "^0.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"